fix(mail): widen thread display
